Refactored departures
