+++ src/styles/modules/_error.scss
@@ -2,7 +2,8 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  padding-top: 10%;
+  padding: 35px 5px 0px 5px;
   background-color: rgb(205, 92, 92);
   color: white;
+  text-align: center;
 }
